backend/pyproject.toml
[project]
name = "whattocook"
version = "0.1.0"
description = "AI-powered recipe generation from fridge photos"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Settings
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.4.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.32",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    # AI / LLM
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    # Redis
    "redis>=5.0.0",
    # Object storage
    "aioboto3>=13.0.0",
    # HTTP client
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whattocook"]

# Pytest and ruff config lives in the root pyproject.toml.
# Keep only a minimal pythonpath hint so running pytest from backend/ still works.
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
strict = true
